requirements.txt
# Core dependencies
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
Pillow>=9.5.0

# Configuration and utilities
PyYAML>=6.0
tqdm>=4.65.0
tensorboard>=2.13.0

# Data processing and augmentation
albumentations>=1.3.0
opencv-python>=4.7.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Development tools
pytest>=7.3.0
flake8>=6.0.0
black>=23.3.0
isort>=5.12.0

# Mixed precision training
apex>=0.1  # Optional, for fp16 training

# Weights & Biases integration (optional)
wandb>=0.15.0 